+++ frontend/src/components/pages/TeamsNew.tsx
@@ -1,51 +1,12 @@
-import { useContext, useState } from "react";
-import { useNavigate } from "react-router-dom";
+import { useState } from "react";
 import { Button, Stack, TextField, Typography } from "@mui/material";
 import Grid2 from "@mui/material/Unstable_Grid2/Grid2";
-import { AxiosRequestConfig, AxiosResponse } from "axios";
-import { baseAxios } from "../../apis/axios";
-import { TeamsResponse } from "../../types/teamTypes";
-import { SnackbarContext } from "../../providers/SnackbarProvider";
 import { BackButton } from "../ui/BackButton";
+import { usePostTeam } from "../../hooks/team/usePostTeam";
 
 export const TeamsNew = () => {
-  const { handleSetSnackbar } = useContext(SnackbarContext);
-  const navigate = useNavigate();
   const [newTeamName, setNewTeamName] = useState<string>("");
-
-  const handleTeamsCreate = () => {
-    const teamsCreateOptions: AxiosRequestConfig = {
-      url: "/teams",
-      method: "POST",
-      data: { name: newTeamName },
-    };
-
-    baseAxios(teamsCreateOptions)
-      .then((res: AxiosResponse<TeamsResponse>) => {
-        console.log(res);
-        handleSetSnackbar({
-          open: true,
-          type: "success",
-          message: "チームを作成しました",
-        });
-        navigate("/sign_up", {
-          state: {
-            teamId: res.data.team.id,
-            teamName: res.data.team.name,
-            isAdmin: true,
-          },
-        });
-      })
-      .catch((err) => {
-        console.log(err);
-        handleSetSnackbar({
-          open: true,
-          type: "error",
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access, @typescript-eslint/no-unsafe-call
-          message: `${err.response.data.messages.join("。")}`,
-        });
-      });
-  };
+  const handleCreateTeam = usePostTeam(newTeamName);
 
   return (
     <Grid2 container direction="column" rowSpacing={3}>
@@ -72,7 +33,7 @@ export const TeamsNew = () => {
       </Grid2>
       <Grid2 xs={12}>
         <Stack direction="row" spacing={1}>
-          <Button type="submit" variant="contained" onClick={handleTeamsCreate}>
+          <Button type="submit" variant="contained" onClick={handleCreateTeam}>
             作成
           </Button>
           <BackButton />
